tests/wpt/tests/workers/Worker_ErrorEvent_bubbles_cancelable.htm

<!DOCTYPE html>
<meta charset="utf-8">
<title>Web Workers: Worker ErrorEvent - bubbles, cancelable</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
// The worker events races with the window's load event; if the worker events
// arrive first, the harness will detect the error event and fail the test.
setup({ allow_uncaught_exception: true });

async_test(function(t) {
  var worker = new Worker('./support/ErrorEvent.js');
  worker.onerror = t.step_func_done(function(e) {
    assert_false(e.bubbles, "onerror on worker doesn't bubble");
    assert_true(e.cancelable, "onerror on worker is cancelable");
  });
  worker.postMessage("Error Message");
}, "ErrorEvent on worker doesn't bubble and is cancelable");

</script>
